refactor: move secondary button
