+++ packages/components/buttons/secondary-button/src/secondary-button.story.js
@@ -3,9 +3,9 @@ import { MemoryRouter, Link } from 'react-router-dom';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { withKnobs, boolean, text, select } from '@storybook/addon-knobs/react';
-import Section from '../../../../.storybook/decorators/section';
-import * as icons from '../../icons';
-import Readme from './README.md';
+import * as icons from '@commercetools-uikit/icons';
+import Section from '../../../../../.storybook/decorators/section';
+import Readme from '../README.md';
 import SecondaryButton from './secondary-button';
 
 const iconNames = Object.keys(icons);
