+++ packages/components/buttons/secondary-button/src/secondary-button.visualroute.js
@@ -3,7 +3,7 @@ import {
   InformationIcon,
   SecondaryButton,
 } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 export const routePath = '/secondary-button';
