+++ src/components/fields/select-field/select-field.form.story.js
@@ -5,10 +5,9 @@ import omitEmpty from 'omit-empty';
 import { action } from '@storybook/addon-actions';
 import { withKnobs, boolean, select } from '@storybook/addon-knobs/react';
 import Spacings from '@commercetools-uikit/spacings';
-import PrimaryButton from '@commercetools-uikit/primary-button';
+import { PrimaryButton, SecondaryButton } from '@commercetools-uikit/buttons';
 import Section from '../../../../.storybook/decorators/section';
 import FormikBox from '../../../../.storybook/decorators/formik-box';
-import SecondaryButton from '../../buttons/secondary-button';
 import Readme from './README.md';
 import SelectField from './select-field';
