+++ src/components/inputs/async-creatable-select-input/async-creatable-select-input.form.story.js
@@ -5,10 +5,9 @@ import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { withKnobs, boolean, number } from '@storybook/addon-knobs/react';
 import Spacings from '@commercetools-uikit/spacings';
-import PrimaryButton from '@commercetools-uikit/primary-button';
+import { PrimaryButton, SecondaryButton } from '@commercetools-uikit/buttons';
 import Section from '../../../../.storybook/decorators/section';
 import FormikBox from '../../../../.storybook/decorators/formik-box';
-import SecondaryButton from '../../buttons/secondary-button';
 import ErrorMessage from '../../messages/error-message';
 import Readme from './README.md';
 import AsyncCreatableSelectInput from './async-creatable-select-input';
