+++ src/components/inputs/select-input/select-input.form.story.js
@@ -4,11 +4,10 @@ import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { withKnobs, boolean } from '@storybook/addon-knobs/react';
 import Spacings from '@commercetools-uikit/spacings';
-import PrimaryButton from '@commercetools-uikit/primary-button';
+import { PrimaryButton, SecondaryButton } from '@commercetools-uikit/buttons';
 import Section from '../../../../.storybook/decorators/section';
 import FormikBox from '../../../../.storybook/decorators/formik-box';
 import ErrorMessage from '../../messages/error-message';
-import SecondaryButton from '../../buttons/secondary-button';
 import Readme from './README.md';
 import SelectInput from './select-input';
