+++ src/components/tooltip/tooltip.story.js
@@ -9,7 +9,7 @@ import {
   select,
   number,
 } from '@storybook/addon-knobs/react';
-import PrimaryButton from '@commercetools-uikit/primary-button';
+import { PrimaryButton } from '@commercetools-uikit/buttons';
 import Section from '../../../.storybook/decorators/section';
 import Readme from './README.md';
 import Tooltip from './tooltip';
